more alpha compile fixes. Closes: #60 because while it started out for 0.1.1 it developed into an alpha compile issue for 0.5.2

+++ libs/video/renderer/sw32/d_edge.c
@@ -202,7 +202,7 @@ D_DrawSurfaces (void)
 			d_zistepv = s->d_zistepv;
 			d_ziorigin = s->d_ziorigin;
 
-			D_DrawSolidSurface (s, ((int) s->data & 0xFF));
+			D_DrawSolidSurface (s, ((long) s->data & 0xFF));
 			D_DrawZSpans (s->spans);
 		}
 	} else {
