+++ libs/video/renderer/sw32/sw32_rmisc.c
@@ -258,8 +258,9 @@ R_SetupFrame (void)
 		if ((surface_p - surfaces) > r_maxsurfsseen)
 			r_maxsurfsseen = surface_p - surfaces;
 
-		Con_Printf ("Used %d of %d surfs; %d max\n", surface_p - surfaces,
-					surf_max - surfaces, r_maxsurfsseen);
+		Con_Printf ("Used %ld of %ld surfs; %d max\n",
+					(long) (surface_p - surfaces),
+					(long) (surf_max - surfaces), r_maxsurfsseen);
 	}
 
 	if (r_numedges->int_val) {
